Add files via upload
This is story 3. Implementing where a user is able to choose a starting city, and based on their choice of city it will show them the list of cities to go to and the total distance traveled. It will also display where they start from. 

I also took the time to organize things from story1.h Do not know if you want to use this implementation, but I thought it was helpful for me. We can also do further organizing later on. Just to help with readability and finding things.